feat(communication): add server capabilities with GUI
Added a server capability to the ESP32 microcontroller, now it is possible to enter the local IP of the controller and make changes. The GUI includes 3 buttons, warm for warm color, cold for cold color and auto for circadian automation. It also included a brightness slider and a button to set the brightness for brightness control.