refactor(all): separate html and c++ code
Separate Client and Server code, installed LittleFS file system to upload files directly to the ESP32 board. Now the html code is no longer strings inside the .ino file and a separate html file on it's own, making further development to the gui easier without having to compile and upload the sketch every time. Also introduced a new Presets section which is still a wip.