wip(structure): migrate project to client server architecture
Create new server for more esp32 microcontrollers, aim is to have 2 clients and 1 server for now, clients handle the rgb strips, server handles the NTP requests, my aim is to have them communicate in realtime, making the clients subscibers to the time received by the server and also have the server control the lights. basically a user logs into the gui through the server, presses a button and the server sends instructions to both client esp32 micros.